pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "opsagent-mcp"
version = "0.1.0"
description = "AI-powered CI/CD failure first-responder using LangGraph + MCP"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "langgraph>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-mcp-adapters>=0.1.0",
    "fastmcp>=2.0.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
build   = ["pyinstaller>=6.0"]
openai  = ["langchain-openai>=0.2.0"]
google  = ["langchain-google-genai>=2.0.0"]
all-providers = ["langchain-openai>=0.2.0", "langchain-google-genai>=2.0.0"]

[project.scripts]
opsagent = "cli:main"

[tool.hatch.build.targets.wheel]
packages = ["."]
